Initial commit, import from Wikisource
There seems to be one single transcription of the book in various forms.
I cannot figure out which edition of the book it comes from, as the ones
on the internet archive are from the late 1800s, and they don't contain
the dedication (in fact, the dedication seems to have been morphed into
a preface). These are the ones I know of:

- Gutenberg[1]/Gutenberg Aus[2]. Very old, contains lots of stars (*)
  for <hr/>s, many instances of <pre> and <br/>s. No italics or bold
  text.
- ebooks@Adelaide[3]. Seems to have been generated from a master copy.
  Better typogrify, but still has stars. No italics or bold text.
- Wikisource[4]. Has stars and divs, but should be easier to remove.
  Also contains italics and most bold text, I did notice some caps.

The only changes I made from the source epub was to enclose with the
standard se template and proper heading blocks.

[1] https://www.gutenberg.org/ebooks/3424
[2] http://gutenberg.net.au/ebooks/e00016.html
[3] https://web.archive.org/web/20190409113110if_/https://ebooks.adelaide.edu.au/cgi-bin/zip/c/clarke/marcus/c59f
[4] https://en.wikisource.org/wiki/For_the_Term_of_His_Natural_Life